e2e: dedupe web-search mock URL and fix fixture doc placement
Collapse the duplicated web-search mock URL into a single source of truth
(WEB_SEARCH_MOCK_NETWORK_URL) and co-locate WEB_SEARCH_PLUGIN_CONFIG with the
mock that owns the network identity. Drop the dead getNetworkApiUrl helper and
move the native-citation doc comment back above buildCitationResponse.

Behavior-preserving: resulting plugin config and fixtures are identical.
